conversiones listas: litros a litros litros a mililitros litros a microlitros
gramos a gramos
gramos a miligramos
gramos a microgramos'  V9.5.0

+++ Backend/src/Users/Models/Recursos.js
@@ -29,7 +29,7 @@ const Recurso = sequelize.define('recursos', {
         type: DataTypes.STRING
     },
     capacidad_r: {
-        type: DataTypes.FLOAT
+        type: DataTypes.DECIMAL
     },
     lote: {
         type: DataTypes.STRING
